requirements.txt
cycler==0.10.0
decorator==4.4.2
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.21.1
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
six==1.16.0
